refactoring: Rename to camelCase
