+++ lib/buildConfig.ts
