+++ lib/voteCfpLambda.ts
@@ -2,7 +2,7 @@ import { Construct } from 'constructs'
 import { Stack, StackProps } from 'aws-cdk-lib'
 import { PolicyStatement } from 'aws-cdk-lib/aws-iam'
 import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs'
-import { BuildConfig } from './build-config'
+import { BuildConfig } from './buildConfig'
 import { IFunction } from 'aws-cdk-lib/aws-lambda'
 import { StatefulStack, tableNameMap } from './statefulStack'
